* subversion/libsvn_fs_fs/temp_serializer.c   (svn_fs_fs__deserialize_manifest): A few manual casts to silence errors.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1124195 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_fs_fs/temp_serializer.c
@@ -570,8 +570,8 @@ svn_fs_fs__deserialize_manifest(void **out,
 {
   apr_array_header_t *manifest = apr_array_make(pool, 1, sizeof(apr_off_t));
 
-  manifest->nelts = data_len / sizeof(apr_off_t);
-  manifest->nalloc = data_len / sizeof(apr_off_t);
+  manifest->nelts = (int) (data_len / sizeof(apr_off_t));
+  manifest->nalloc = (int) (data_len / sizeof(apr_off_t));
   manifest->elts = (char*)data;
 
   *out = manifest;
